# V2.3/code/_openapi_snapshot.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Starriver SuperBrain API",
    "description": "Minimal runnable V2.3 API skeleton",
    "version": "v2.3-preview"
  },
  "paths": {
...
    "/api/v2.3-preview/consciousness/attention": {
      "get": {
        "tags": [
          "consciousness"
        ],
        "summary": "Get Attention",
        "operationId": "get_attention_api_v2_3_preview_consciousness_attention_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttentionResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "consciousness"
        ],
        "summary": "Set Attention",
        "operationId": "set_attention_api_v2_3_preview_consciousness_attention_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttentionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttentionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2.3-preview/consciousness/state": {
      "get": {
        "tags": [
          "consciousness"
        ],
        "summary": "Get State",
        "operationId": "get_state_api_v2_3_preview_consciousness_state_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StateResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "consciousness"
        ],
        "summary": "Set State",
        "operationId": "set_state_api_v2_3_preview_consciousness_state_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2.3-preview/cloud/consent": {
      "post": {
        "tags": [
          "cloud"
        ],
        "summary": "Create Consent",
        "operationId": "create_consent_api_v2_3_preview_cloud_consent_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsentRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "cloud"
        ],
        "summary": "Revoke Consent",
        "operationId": "revoke_consent_api_v2_3_preview_cloud_consent_delete",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "User ID to revoke consent for",
              "title": "User Id"
            },
            "description": "User ID to revoke consent for"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2.3-preview/cloud/status": {
      "get": {
        "tags": [
          "cloud"
        ],
        "summary": "Consent Status",
        "operationId": "consent_status_api_v2_3_preview_cloud_status_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "User ID to get consent status for",
              "title": "User Id"
            },
            "description": "User ID to get consent status for"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AttentionRequest": {
        "properties": {
          "target": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target"
          },
          "mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mode",
            "default": "push"
          }
        },
        "type": "object",
        "title": "AttentionRequest"
      },
      "AttentionResponse": {
        "properties": {
          "current": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current"
          },
          "stack": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Stack"
          },
          "stack_size": {
            "type": "integer",
            "title": "Stack Size"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "current",
          "stack",
          "stack_size",
          "updated_at"
        ],
        "title": "AttentionResponse"
      },
      "ConsentRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "consent": {
            "type": "boolean",
            "title": "Consent",
            "default": true
          },
          "scopes": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scopes",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "ConsentRequest"
      },
      "ConsentResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "consent_id": {
            "type": "string",
            "title": "Consent Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "granted_scopes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Granted Scopes"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "status",
          "consent_id",
          "user_id",
          "granted_scopes",
          "timestamp"
        ],
        "title": "ConsentResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StateRequest": {
        "properties": {
          "state": {
            "type": "string",
            "title": "State"
          },
          "goal": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Goal"
          },
          "force": {
            "type": "boolean",
            "title": "Force",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "state"
        ],
        "title": "StateRequest"
      },
      "StateResponse": {
        "properties": {
          "state": {
            "type": "string",
            "title": "State"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "current_goal": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Goal"
          },
          "goal_stack": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Goal Stack"
          },
          "allowed_next_states": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allowed Next States"
          }
        },
        "type": "object",
        "required": [
          "state",
          "updated_at",
          "current_goal",
          "goal_stack",
          "allowed_next_states"
        ],
        "title": "StateResponse"
      },
      "StatusResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "consent_active": {
            "type": "boolean",
            "title": "Consent Active"
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Scopes"
          }
        },
        "type": "object",
        "required": [
          "status",
          "consent_active",
          "scopes"
        ],
        "title": "StatusResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}